Move test tools into test.tools
